+++ src/__tests__/FormAutocompleteField.test.tsx
@@ -1,20 +1,18 @@
-//import React from "react"
+import React from "react"
 
 import "@testing-library/jest-dom"
-//import { ThemeProvider, StyledEngineProvider } from "@mui/material/styles"
-//import { render, screen, waitFor, act, fireEvent } from "@testing-library/react"
+import { ThemeProvider, StyledEngineProvider } from "@mui/material/styles"
+import { screen, waitFor, fireEvent } from "@testing-library/react"
 //import userEvent from "@testing-library/user-event"
 import { rest } from "msw"
 import { setupServer } from "msw/node"
-//import { Provider } from "react-redux"
-//import configureStore from "redux-mock-store"
+import { act } from "react-dom/test-utils"
 
-//import WizardFillObjectDetailsForm from "../components/SubmissionWizard/WizardForms/WizardFillObjectDetailsForm"
-//import CSCtheme from "../theme"
+import WizardFillObjectDetailsForm from "../components/SubmissionWizard/WizardForms/WizardFillObjectDetailsForm"
+import CSCtheme from "../theme"
 
-//import { ObjectSubmissionTypes, ObjectTypes } from "constants/wizardObject"
-
-//const mockStore = configureStore([])
+import { ObjectSubmissionTypes, ObjectTypes } from "constants/wizardObject"
+import { renderWithProviders } from "utils/test-utils"
 
 const mockOrganisations = [{ name: "Test Organisation" }, { name: "Mock Org" }]
 
@@ -37,17 +35,21 @@ describe("Test autocomplete on organisation field", () => {
     },
   }
 
-  //const store = mockStore({
-  //  objectType: ObjectTypes.dac,
-  //  submissionType: ObjectSubmissionTypes.form,
-  //  submission: {
-  //    description: "Test desciption",
-  //    id: "FOL90524783",
-  //    name: "Test name",
-  //    published: false,
-  //  },
-  //  openedXMLModal: false,
-  //})
+  const preloadedState = {
+    objectType: ObjectTypes.dac,
+    submissionType: ObjectSubmissionTypes.form,
+    submission: {
+      description: "Test desciption",
+      submissionId: "FOL90524783",
+      name: "Test name",
+      published: false,
+      workflow: "FEGA",
+      drafts: [],
+      metadataObjects: [],
+      doiInfo: { creators: [], contributors: [], subjects: [] },
+    },
+    openedXMLModal: false,
+  }
 
   beforeAll(() => server.listen())
   afterEach(() => server.resetHandlers())
@@ -56,40 +58,43 @@ describe("Test autocomplete on organisation field", () => {
   sessionStorage.setItem(`cached_dac_schema`, JSON.stringify(schema))
 
   test("should render autocomplete field if schema has 'organisation' property", async () => {
-    //jest.setTimeout(30000)
-    //render(
-    //  <Provider store={store}>
-    //    <StyledEngineProvider injectFirst>
-    //      <ThemeProvider theme={CSCtheme}>
-    //        <WizardFillObjectDetailsForm />
-    //      </ThemeProvider>
-    //    </StyledEngineProvider>
-    //  </Provider>
-    //)
-    //await waitFor(() => {
-    //  const autocomplete = screen.getByTestId("organisation-inputField")
-    //  expect(autocomplete).toBeDefined()
-    //})
+    jest.setTimeout(30000)
+    act(() => {
+      renderWithProviders(
+        <StyledEngineProvider injectFirst>
+          <ThemeProvider theme={CSCtheme}>
+            <WizardFillObjectDetailsForm />
+          </ThemeProvider>
+        </StyledEngineProvider>,
+        { preloadedState }
+      )
+    })
+    await waitFor(() => {
+      const autocomplete = screen.getByTestId("organisation-inputField")
+      expect(autocomplete).toBeDefined()
+    })
   })
 
   test("should search for organisations", async () => {
-    //render(
-    //  <Provider store={store}>
-    //    <StyledEngineProvider injectFirst>
-    //      <ThemeProvider theme={CSCtheme}>
-    //        <WizardFillObjectDetailsForm />
-    //      </ThemeProvider>
-    //    </StyledEngineProvider>
-    //  </Provider>
-    //)
-    //const autocomplete = (await waitFor(() => screen.getByTestId("organisation-inputField"))) as HTMLInputElement
-    //autocomplete.focus()
-    //act(() => {
-    //  // Assign value to autocomplete field
-    //  // Note: userEvent doesn't work inside act(), which in this case is needed for MUI autocomplete field
-    //  fireEvent.change(autocomplete, { target: { value: "test" } })
-    //})
-    //// Find loading indicator
+    act(() => {
+      renderWithProviders(
+        <StyledEngineProvider injectFirst>
+          <ThemeProvider theme={CSCtheme}>
+            <WizardFillObjectDetailsForm />
+          </ThemeProvider>
+        </StyledEngineProvider>,
+        { preloadedState }
+      )
+    })
+    const autocomplete = (await waitFor(() => screen.getByTestId("organisation-inputField"))) as HTMLInputElement
+
+    act(() => {
+      autocomplete.focus()
+      // Assign value to autocomplete field
+      // Note: userEvent doesn't work inside act(), which in this case is needed for MUI autocomplete field
+      fireEvent.change(autocomplete, { target: { value: "test" } })
+    })
+    // Find loading indicator
     //await waitFor(() => screen.getByRole("progressbar"))
     //// Find options wrapper
     //await waitFor(() => screen.getByRole("presentation"))