+++ src/__tests__/WizardStepper.test.tsx
@@ -1,92 +1,89 @@
-//import React from "react"
+import React from "react"
 
 import "@testing-library/jest-dom"
-//import { ThemeProvider, StyledEngineProvider } from "@mui/material/styles"
-//import { render, screen, fireEvent } from "@testing-library/react"
-//import { Provider } from "react-redux"
-//import { MemoryRouter, Routes, Route } from "react-router-dom"
-//import configureMockStore from "redux-mock-store"
-//import { thunk } from "redux-thunk"
+import { ThemeProvider, StyledEngineProvider } from "@mui/material/styles"
+import { screen } from "@testing-library/react"
+import { MemoryRouter, Routes, Route } from "react-router-dom"
 
-//import WizardStepper from "../components/SubmissionWizard/WizardComponents/WizardStepper"
-//import CSCtheme from "../theme"
+import WizardStepper from "../components/SubmissionWizard/WizardComponents/WizardStepper"
+import CSCtheme from "../theme"
 
-//import { ObjectSubmissionTypes, ObjectTypes } from "constants/wizardObject"
-
-//const middlewares = []
-//const mockStore = configureMockStore(middlewares)
+import { ObjectSubmissionTypes, ObjectTypes } from "constants/wizardObject"
+import { Schema } from "types"
+import { renderWithProviders } from "utils/test-utils"
 
 describe("WizardStepper", () => {
   it("should disable next steps when naming submission", () => {
-    //const store = mockStore({
-    //  submissionType: "",
-    //  objectTypesArray: Object.keys(ObjectTypes),
-    //  submission: {
-    //    submissionId: "",
-    //    name: "",
-    //    description: "",
-    //    published: false,
-    //    drafts: [],
-    //    metadataObjects: [],
-    //    doiInfo: { creators: [], contributors: [], subjects: [] },
-    //  },
-    //  stepObject: { step: 1, stepObjectType: "submissionDetails" },
-    //})
-    //render(
-    //  <MemoryRouter initialEntries={[{ pathname: "/submission", search: "?step=1" }]}>
-    //    <Provider store={store}>
-    //      <Routes>
-    //        <Route
-    //          path="/submission"
-    //          element={
-    //            <StyledEngineProvider injectFirst>
-    //              <ThemeProvider theme={CSCtheme}>
-    //                <WizardStepper />
-    //              </ThemeProvider>
-    //            </StyledEngineProvider>
-    //          }
-    //        />
-    //      </Routes>
-    //    </Provider>
-    //  </MemoryRouter>
-    //)
-    //const secondStep = screen.getByTestId("2-step-disabled")
-    //expect(secondStep).toBeInTheDocument()
+    renderWithProviders(
+      <MemoryRouter initialEntries={[{ pathname: "/submission", search: "?step=1" }]}>
+        <Routes>
+          <Route
+            path="/submission"
+            element={
+              <StyledEngineProvider injectFirst>
+                <ThemeProvider theme={CSCtheme}>
+                  <WizardStepper />
+                </ThemeProvider>
+              </StyledEngineProvider>
+            }
+          />
+        </Routes>
+      </MemoryRouter>,
+      {
+        preloadedState: {
+          submissionType: "",
+          objectTypesArray: Object.keys(ObjectTypes) as Schema[],
+          submission: {
+            submissionId: "",
+            name: "",
+            description: "",
+            published: false,
+            drafts: [],
+            metadataObjects: [],
+            workflow: "",
+            doiInfo: { creators: [], contributors: [], subjects: [] },
+          },
+          stepObject: { step: 1, stepObjectType: "submissionDetails" },
+        },
+      }
+    )
+    expect(screen.queryByTestId("2-step-disabled")).toBeNull()
   })
   it("should open dialog if form or upload in progress", () => {
-    //const store = mockStore({
-    //  submissionType: ObjectSubmissionTypes.form,
-    //  objectTypesArray: Object.keys(ObjectTypes),
-    //  submission: {
-    //    submissionId: "FOLD1234",
-    //    name: "Test submission",
-    //    description: "Submission description",
-    //    published: false,
-    //    drafts: [],
-    //    metadataObjects: [],
-    //    doiInfo: { creators: [], contributors: [], subjects: [] },
-    //  },
-    //  stepObject: { step: 2, stepObjectType: "study" },
-    //  draftStatus: "notSaved",
-    //})
-    //render(
-    //  <MemoryRouter initialEntries={[{ pathname: "/submission", search: "?step=2" }]}>
-    //    <Provider store={store}>
-    //      <Routes>
-    //        <Route
-    //          path="/submission"
-    //          element={
-    //            <StyledEngineProvider injectFirst>
-    //              <ThemeProvider theme={CSCtheme}>
-    //                <WizardStepper />
-    //              </ThemeProvider>
-    //            </StyledEngineProvider>
-    //          }
-    //        />
-    //      </Routes>
-    //    </Provider>
-    //  </MemoryRouter>
-    //)
+    renderWithProviders(
+      <MemoryRouter initialEntries={[{ pathname: "/submission", search: "?step=2" }]}>
+        <Routes>
+          <Route
+            path="/submission"
+            element={
+              <StyledEngineProvider injectFirst>
+                <ThemeProvider theme={CSCtheme}>
+                  <WizardStepper />
+                </ThemeProvider>
+              </StyledEngineProvider>
+            }
+          />
+        </Routes>
+      </MemoryRouter>,
+      {
+        preloadedState: {
+          submissionType: ObjectSubmissionTypes.form,
+          objectTypesArray: Object.keys(ObjectTypes) as Schema[],
+          submission: {
+            submissionId: "FOLD1234",
+            name: "Test submission",
+            description: "Submission description",
+            published: false,
+            drafts: [],
+            metadataObjects: [],
+            workflow: "",
+            doiInfo: { creators: [], contributors: [], subjects: [] },
+          },
+          stepObject: { step: 2, stepObjectType: "study" },
+          draftStatus: "notSaved",
+        },
+      }
+    )
     //const button = screen.getByRole("button", { name: "Add DAC" })
     //fireEvent.click(button)
     //expect(screen.getByRole("dialog")).toBeDefined